+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 * `Client.AddressableDevices` returns cached value, no errors returned.
 * Added `BufferFromHost.ToDeviceNum` to allow specification of the device by device number in the addressable devices list. 
 * Added `LoadedExecutable.Execute.OnDeviceNum` to allow  specification of the device by device number in the addressable devices list.
+* Removed the awkward `pjrt.FlatDataToRawWithDimensions` and added the more ergonomic `Client.BufferFromHost.FromFlatDataWithDimensions`.
 
 # v0.1.2 SuppressAbseilLoggingHack